requirements.txt
Django==1.10
PyMySQL==0.7.8
Pillow==3.3.0
django-imagekit==3.3
django-modeldict==1.4.1
django-meta==1.3.1
uwsgi==2.0.13.1
python-amazon-simple-product-api==2.1.0
djangorestframework==3.4.6
beautifulsoup4==4.5.1